+++ packages/core/src/api/factories/contract/getAllContractIds.ts
@@ -3,6 +3,7 @@
  */
 import {ChainService} from '../../../service/chainService';
 import {ContractIdList} from '../../../typings/contractIdList';
+import {GetAllContractIdsArgs} from '../../../typings/args/getAllContractIdsArgs';
 
 /**
  * Use with [[ApiComposer]] and belongs to [[ContractApi]].
@@ -11,6 +12,11 @@ import {ContractIdList} from '../../../typings/contractIdList';
  * @module core.api.factories
  */
 export const getAllContractIds = (service: ChainService):
-    () => Promise<ContractIdList> =>
-    (): Promise<ContractIdList> =>
-        service.query('getATIds');
+    (args: GetAllContractIdsArgs) => Promise<ContractIdList> =>
+    (args: GetAllContractIdsArgs): Promise<ContractIdList> => {
+        const params = {
+            machineCodeHashId: args.machineCodeHash
+        };
+
+        return service.query('getATIds', params);
+    };