+++ packages/core/src/api/factories/contract/getContractsByAccount.ts
@@ -1,8 +1,10 @@
 /**
  * Copyright (c) 2019 Burst Apps Team
+ * Modified (c) 2021 Signum Network
  */
 import {ChainService} from '../../../service/chainService';
 import {ContractList} from '../../../typings/contractList';
+import {GetContractsByAccountArgs} from '../../../typings/args/getContractsByAccountArgs';
 
 /**
  * Use with [[ApiComposer]] and belongs to [[ContractApi]].
@@ -11,8 +13,11 @@ import {ContractList} from '../../../typings/contractList';
  * @module core.api.factories
  */
 export const getContractsByAccount = (service: ChainService):
-    (accountId: string) => Promise<ContractList> =>
-    (accountId: string): Promise<ContractList> =>
-        service.query('getAccountATs', {
-            account: accountId,
-        });
+    (args: GetContractsByAccountArgs) => Promise<ContractList> =>
+    (args: GetContractsByAccountArgs): Promise<ContractList> => {
+        const params = {
+            account: args.accountId,
+            machineCodeHashId: args.machineCodeHash
+        };
+        return service.query('getAccountATs', params);
+    };