+++ packages/core/src/typings/api/contractApi.ts
@@ -1,9 +1,11 @@
 import {Contract} from '../contract';
 import {ContractList} from '../contractList';
 import {ContractIdList} from '../contractIdList';
-import {PublishContractArgs} from '../args';
+import {PublishContractArgs, PublishContractByReferenceArgs} from '../args';
 import {TransactionId} from '../transactionId';
 import {CallContractMethodArgs} from '../args/callContractMethodArgs';
+import {GetAllContractIdsArgs} from '../args/getAllContractIdsArgs';
+import {GetContractsByAccountArgs} from '../args/getContractsByAccountArgs';
 
 /**
  * Contract API
@@ -21,18 +23,19 @@ export interface ContractApi {
     getContract: (id: string) => Promise<Contract>;
 
     /**
-     * Get all contracts of given account
-     * @param accountId The account id (not RS address)
+     * Get all contracts of given account, or filtered by the arguments
+     * @param args {GetContractsByAccountArgs} The argument object
      * @returns {ContractList} A list of contracts
      */
-    getContractsByAccount: (accountId: string) => Promise<ContractList>;
+    getContractsByAccount: (args: GetContractsByAccountArgs) => Promise<ContractList>;
 
 
     /**
-     * Get all contract Ids of the blockchain
+     * Get all contract Ids of the blockchain, or filtered by the argument
+     * @param args {GetAllContractIdsArgs} The argument object
      * @return {ContractIdList} The list of contract ids
      */
-    getAllContractIds: (id: string) => Promise<ContractIdList>;
+    getAllContractIds: (args: GetAllContractIdsArgs) => Promise<ContractIdList>;
 
 
     /**
@@ -41,6 +44,14 @@ export interface ContractApi {
      * @return {TransactionId} The transaction id in case of success
      */
     publishContract: (args: PublishContractArgs) => Promise<TransactionId>;
+
+   /**
+     * Publishes a smart contract to the blockchain using the code base of an already existant contract.
+     * This reduces payload in the chain, allowing to deploy many contracts per block and very cheap
+     * @param args {PublishContractArgs} The argument object
+     * @return {TransactionId} The transaction id in case of success
+     */
+    publishContractByReference: (args: PublishContractByReferenceArgs) => Promise<TransactionId>;
 
     /**
      * Calls a public method of smart contract