+++ packages/core/src/typings/args/getAllContractIdsArgs.ts
@@ -1,0 +1,12 @@
+
+/**
+ * The argument object for [[ContractApi.getAllContractIds]]
+ *
+ * @module core
+ */
+export interface GetAllContractIdsArgs {
+    /**
+     * The hash of the machine code
+     */
+    machineCodeHash?: string;
+}