+++ packages/core/CHANGELOG.md
@@ -8,8 +8,10 @@ All notable changes to this project will be documented in this file.
 - Rebranding
 - Removed all deprecated functions
 - Several Renamed Functions and classes
+- Support for Contract References aka Carbon Contracts/Green Smart Contracts
 - New Functions:
-  - getMiningInfo
+  - `getMiningInfo`
+  - `publishContractByReference`
 - Minor bug fixes
 
 ----------------------------------------------------